test scoped based privileges
