+++ data-attribute-resolver.d.ts
@@ -1,0 +1,13 @@
+import { QueryEntity, QueryField } from '@themost/query';
+
+export declare class DataAttributeResolver {
+    orderByNestedAttribute(attr: string): any;
+    selectNestedAttribute(attr: string): any;
+    selectAggregatedAttribute(aggregation: string, attribute: string, alias: string): any;
+    resolveNestedAttribute(attr: string): any;
+    resolveNestedAttributeJoin(memberExpr: string): { $select?: QueryField, $expand?: QueryEntity[] };
+    testAttribute(s: string): any;
+    testAggregatedNestedAttribute(s: string): any;
+    testNestedAttribute(s: string): any;
+    resolveJunctionAttributeJoin(attr: string): { $select?: QueryField, $expand?: QueryEntity[] };
+}
