+++ spec/test2/config/models/AnotherStructuredValue.json
@@ -1,0 +1,14 @@
+{
+    "$schema": "https://themost-framework.github.io/themost/models/2018/2/schema.json",
+    "@id": "http://schema.org/AnotherStructuredValue",
+    "name": "AnotherStructuredValue",
+    "description": "Structured values are used when the value of a property has a more complex structure than simply being a textual value or a reference to another thing.",
+    "title": "AnotherStructuredValue",
+    "abstract": false,
+    "sealed": false,
+    "inherits": "Intangible",
+    "version": "2.0",
+    "fields": [],
+    "privileges": [
+    ]
+}
